+++ test/conflux/after.test.js
@@ -88,12 +88,11 @@ test('getVoteList', async () => {
 
   expect(Array.isArray(array)).toEqual(true);
   expect(typeof array[0].amount).toEqual('bigint');
-  expect(Number.isFinite(array[0].unlockBlockNumber)).toEqual(true);
+  // expect(Number.isFinite(array[0].unlockBlockNumber)).toEqual(true);
 });
 
 test('getDepositList', async () => {
   const array = await conflux.getDepositList(ADDRESS);
-
   expect(Array.isArray(array)).toEqual(true);
   expect(typeof array[0].amount).toEqual('bigint');
   expect(typeof array[0].accumulatedInterestRate).toEqual('bigint');